filters/ruby: Modified the ruby filter and made it more robust.
Previously, if the ruby filter encountered an exception, it would silently
fail. Not only would the filter fail, but it would take the entire logstash
process down with it (due to an unhandled exception).

The changes herein add error checking (to prevent unhandled exceptions from
slipping through and crashing logstash) and report any errors encountered.